161 task date time (#162)
* Task due date rules & v-chip with changing colors

* Deleting task now requires confirm

* ViewCourses now with auto width, looking like the other Views

+++ modules/core/web/src/dialog/DialogEditTask.vue
@@ -10,7 +10,7 @@
             <v-select v-model="currentTask.solutionModelId" label="Musterdiagram" variant="underlined" :items="diagrams" item-title="name" item-value="id" color="#81ba24"></v-select>
           </div>
           <div>
-            <v-text-field v-model="currentTask.dueDate" label="Deadline" variant="underlined" color="#81ba24"></v-text-field>
+            <v-text-field v-model="currentTask.dueDate" :rules="dueDateRules" label="Deadline" variant="underlined" color="#81ba24" hint="DD.MM.YYYY HH:MM"></v-text-field>
             <v-select v-if="!workInProgress" v-model="currentTask.mediaType" :items="['Model', 'Text']" label="Medientyp" variant="underlined" required color="#81ba24"></v-select>
             <v-select v-if="!workInProgress" v-model="currentTask.rulesetId" :items="rulesets" item-title="name" label="Regelsatz" variant="underlined" required color="#81ba24" item-value="id"></v-select>
             <v-select v-model="currentTask.eliability" :items="liabilities" label="Verpflichtung" variant="underlined" required color="#81ba24" item-title="name" item-value="enum"></v-select>
@@ -27,6 +27,7 @@
     </v-card>
     <v-snackbar v-model="snackbarFail" :timeout="3000"> Ein Fehler ist aufgetreten, bitte versuchen Sie es erneut </v-snackbar>
   </v-dialog>
+  <DialogConfirmVue ref="dialogConfirm"></DialogConfirmVue>
 </template>
 
 <script setup lang="ts">
@@ -39,6 +40,7 @@ import { useAuthUserStore } from "@/stores/authUserStore";
 import { useRoute } from "vue-router";
 import taskService from "@/services/task.service";
 import router from "@/router";
+import DialogConfirmVue from "../dialog/DialogConfirm.vue";
 
 const arrayMaxSubmissions = Array.from(Array(100).keys());
 
@@ -48,6 +50,7 @@ const authUserStore = useAuthUserStore();
 const route = useRoute();
 
 const editTaskDialog = ref<boolean>(false);
+const dialogConfirm = ref<typeof DialogConfirmVue>();
 
 const snackbarFail = ref(false);
 const editTitle = ref<string>("");
@@ -73,6 +76,8 @@ const selectedCategoryId = ref<number>();
 const selectedDiagramId = ref<number>();
 const nameRules = ref<any>([(v: string) => !!v || "Name ist erforderlich"]);
 const descriptionRules = ref<any>([(v: string) => !!v || "Beschreibung ist erforderlich"]);
+const regex = /^([0-2][0-9]|3[0-1])\.(0[1-9]|1[0-2])\.\d{4} ([01][0-9]|2[0-3]):[0-5][0-9]$/;
+const dueDateRules = ref<any>([(v: string) => !v || regex.test(v) || "Ungültiges Datum dd.mm.yyyy hh:mm"]);
 
 // empty, or should be a valid date and in the future
 // const dueDateRules = ref<any>([(v: string) => !v || (new Date(v) > new Date() && !isNaN(new Date(v).getTime())) || "Ungültiges Datum"]);
@@ -167,8 +172,12 @@ const _cancel = () => {
 };
 
 const deleteTask = () => {
-  taskService.deleteTask(currentTask.value.id).then(() => {
-    router.push("/course/" + currentTask.value.courseId);
+  dialogConfirm.value?.openDialog(`Lösche Aufgabe`, "Willst du die Aufgabe wirklich löschen?").then((result: boolean) => {
+    if (result) {
+      taskService.deleteTask(currentTask.value.id).then(() => {
+        router.push("/course/" + currentTask.value.courseId);
+      });
+    }
   });
 };
